# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: e-IRF Electronic Incident Records Form API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /incidents:
    get:
      operationId: listIncidents
      tags: [incidents]
      summary: List incidents
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            $ref: "#/components/schemas/IncidentType"
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/IncidentStatus"
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of incidents
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IncidentListResponse"
    post:
      operationId: createIncident
      tags: [incidents]
      summary: Create incident
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IncidentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Incident"
  /incidents/{id}:
    get:
      operationId: getIncident
      tags: [incidents]
      summary: Get incident by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Incident found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Incident"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateIncident
      tags: [incidents]
      summary: Update incident
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IncidentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Incident"
    delete:
      operationId: deleteIncident
      tags: [incidents]
      summary: Delete incident
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /officers:
    get:
      operationId: listOfficers
      tags: [officers]
      summary: List all officers
      responses:
        "200":
          description: Officers list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Officer"
    post:
      operationId: createOfficer
      tags: [officers]
      summary: Create officer account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OfficerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Officer"
  /officers/{id}:
    get:
      operationId: getOfficer
      tags: [officers]
      summary: Get officer by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Officer found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Officer"
    patch:
      operationId: updateOfficer
      tags: [officers]
      summary: Update officer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OfficerUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Officer"
    delete:
      operationId: deleteOfficer
      tags: [officers]
      summary: Delete officer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
...
    OfficerRole:
      type: string
      enum: [admin, officer]
    OfficerInputRole:
      type: string
      enum: [admin, officer]
    OfficerUpdateRole:
      type: string
      enum: [admin, officer]
    Officer:
      type: object
      required: [id, name, badgeNumber, rank, role, username, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        badgeNumber:
          type: string
        rank:
          type: string
        role:
          $ref: "#/components/schemas/OfficerRole"
        username:
          type: string
        createdAt:
          type: string
    OfficerInput:
      type: object
      required: [name, badgeNumber, rank, username, password, role]
      properties:
        name:
          type: string
        badgeNumber:
          type: string
        rank:
          type: string
        username:
          type: string
        password:
          type: string
        role:
          $ref: "#/components/schemas/OfficerInputRole"
    OfficerUpdate:
      type: object
      properties:
        name:
          type: string
        badgeNumber:
          type: string
        rank:
          type: string
        username:
          type: string
        password:
          type: ["string", "null"]
        role:
          $ref: "#/components/schemas/OfficerUpdateRole"
    IncidentStatus:
      type: string
      enum: [open, under_investigation, closed, archived]

    # Was previously a plain `string` on every incident schema below — any
    # caller hitting the API directly (not just the UI's dropdown) could
    # set an arbitrary value, silently fragmenting the "by type" dashboard
    # breakdown and making the record unfindable via the type filter.
    # See U4/B4 in the audit. Keep in sync with
    # lib/db/src/schema/incidents.ts's incidentTypeEnum and
    # artifacts/eirf/src/lib/incident-types.ts.
    IncidentType:
      type: string
      enum: [Crime, Accident, Dispute, "Missing Person", Other]
    Incident:
      type: object
      required: [id, incidentNumber, date, time, location, type, description, status, createdAt, updatedAt]
      properties:
        id:
          type: integer
        incidentNumber:
          type: string
        date:
          type: string
        time:
          type: string
        location:
          type: string
        type:
          $ref: "#/components/schemas/IncidentType"
        description:
          type: string
        status:
          $ref: "#/components/schemas/IncidentStatus"
        reportingOfficerId:
          type: ["integer", "null"]
        reportingOfficerName:
          type: ["string", "null"]
        witnessStatements:
          type: ["string", "null"]
        evidence:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    IncidentInput:
      type: object
      required: [date, time, location, type, description]
      properties:
        date:
          type: string
        time:
          type: string
        location:
          type: string
        type:
          $ref: "#/components/schemas/IncidentType"
        description:
          type: string
        witnessStatements:
          type: string
        evidence:
          type: string
        notes:
          type: string
        status:
          $ref: "#/components/schemas/IncidentStatus"
    IncidentUpdate:
      type: object
      properties:
        date:
          type: string
        time:
          type: string
        location:
          type: string
        type:
          $ref: "#/components/schemas/IncidentType"
        description:
          type: string
        witnessStatements:
          type: ["string", "null"]
        evidence:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        status:
          $ref: "#/components/schemas/IncidentStatus"
    IncidentListResponse:
      type: object
      required: [incidents, total, page, limit]
      properties:
        incidents:
          type: array
          items:
            $ref: "#/components/schemas/Incident"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...